Unify cloud-next routes into cloud: route new sessions to next-gen UI
Remove separate /cloud-next routes and sidebar entries. The /cloud
routes now serve CloudNextSessionsPage directly and use isNewSession()
to decide whether to render the next-gen or legacy chat wrapper.
