--- src/app/(app)/cloud-next/chat/page.tsx
@@ -1,5 +1,0 @@
-import { CloudChatPageWrapper } from './CloudChatPageWrapper';
-
-export default function PersonalCloudNextChatPage() {
-  return <CloudChatPageWrapper />;
-}
