+++ src/app/(app)/cloud/chat/CloudChatPageWrapperNext.tsx
@@ -4,17 +4,13 @@ import { Suspense } from 'react';
 import { CloudAgentProvider } from '@/components/cloud-agent-next/CloudAgentProvider';
 import { CloudChatPage } from '@/components/cloud-agent-next/CloudChatPage';
 
-type CloudChatPageWrapperProps = {
-  organizationId: string;
-};
-
-export function CloudChatPageWrapper({ organizationId }: CloudChatPageWrapperProps) {
+export function CloudChatPageWrapperNext() {
   return (
     <CloudAgentProvider>
       <Suspense
         fallback={<div className="flex h-screen items-center justify-center">Loading...</div>}
       >
-        <CloudChatPage organizationId={organizationId} />
+        <CloudChatPage />
       </Suspense>
     </CloudAgentProvider>
   );
