+++ src/lib/cloud-agent/session-type.ts
@@ -1,0 +1,3 @@
+export function isNewSession(sessionId: string): boolean {
+  return sessionId.startsWith('ses_');
+}
